+++ rewrite/changelog.md
@@ -19,6 +19,8 @@ The subsections are formatted as they need when publishing a release to the Play
 
 ## Unreleased
 
+ * Added audio focus callbacks
+
 ## [0.9.4] - 2019-08-15
 
  * Added wakelock and wifilock